emccリンク成功: Common/SboSockLib含む全314ファイルの非Windowsコンパイル・リンク通過
- PlatformDefs.h: ULONGLONG/HDWP型、ファイルAPI定数、NMHDR構造体、GUID比較演算子、min/max/CRect追加
- Win32ApiStubs.h: CreateFile/ReadFile/WriteFile/CloseHandle等のfopen系スタブ、ウィンドウ操作スタブ追加
- CStringCompat.h: 異なる文字型間の変換コンストラクタ追加、min/maxマクロ干渉回避
- TCharCompat.h: C++20でdeprecatedのwstring_convert/codecvt_utf8を手動UTF-8変換に置換
- SboSockLib/StdAfx.h: 独自型定義をPlatformDefs.h+CStringCompat.hに統一
- AflMusic: mmsystem.h/DirectMusicガード、スレッド関数をSDL2 APIに移行
- 個別ファイル: #ifdefガード追加、narrowing修正、goto変数初期化修正

+++ SboCli/src/StateProc/StateProcLOGINMENU.cpp
@@ -323,7 +323,7 @@ void CStateProcLOGINMENU::OnWindowMsgSTYLESELECT(DWORD dwPara)
 		m_pWindowNAMEINPUT->m_pInfoCharCli->MakeCharGrp();
 		break;
 
-	case -1:
+	case (DWORD)-1:
 		// キャンセルされたので前の画面に戻る
 		m_pMgrWindow->Delete(WINDOWTYPE_STYLESELECT);
 		m_pMgrData->PostWindowMessage(WINDOWTYPE_LOGINMENU, 1);
@@ -342,7 +342,7 @@ void CStateProcLOGINMENU::OnWindowMsgNAMEINPUT(DWORD dwPara)
 		m_pSock->Send(&Packet);
 		break;
 
-	case -1:
+	case (DWORD)-1:
 		// キャンセルされたので前の画面に戻る
 		m_pMgrWindow->Delete(WINDOWTYPE_NAMEINPUT);
 		m_pMgrData->PostWindowMessage(WINDOWTYPE_FAMILYTYPE, 0);
